EmailJS wrapper, loading FakeStore y feedback UI
